+++ microcontroller-src/kv4p_ht_esp32_wroom_32/globals.h
@@ -34,6 +34,7 @@ enum RfModuleType : uint8_t {
 #define AUDIO_FRAME_SAMPLES_48K 747
 #define AUDIO_FRAME_BYTES 128
 #define AUDIO_RESAMPLE_RATIO 3
+#define AUDIO_DSP_ALIGN16 __attribute__((aligned(16)))
 
 inline uint32_t bluetoothDeviceId() {
   uint64_t mac = ESP.getEfuseMac();
@@ -92,9 +93,6 @@ Mode mode = MODE_STOPPED;
 
 // Current SQ status
 bool squelched = false;
-bool hardwareSquelched = false;
-bool softSquelchEnabled = true;
-uint8_t softSquelchDeadbandLevel = 0;
 
 // Forward declarations
 void setMode(Mode newMode);
