+++ microcontroller-src/kv4p_ht_esp32_wroom_32/txAudio.h
@@ -24,7 +24,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <AfskModulator.h>
 #include "globals.h"
 #include "protocol.h"
-#include "audioResampler.h"
+#include "dsp/audioResampler.h"
 
 bool txStreamConfigured = false;
 bool txDecodeStreamStarted = false;
